+++ .travis.yml
@@ -6,8 +6,8 @@ android:
   components:
     - tools
     - platform-tools
-    - build-tools-28.0.3
-    - android-28
+    - build-tools-29.0.3
+    - android-29
     - extra-android-m2repository
     - extra-google-m2repository
     - extra-android-support
