+++ lib/build.gradle
@@ -86,7 +86,7 @@ dependencies {
     implementation 'com.google.android.gms:play-services-maps:17.0.0'
     implementation 'com.google.android.gms:play-services-location:17.0.0'
     // Google maps util
-    implementation 'com.google.maps.android:android-maps-utils:2.0.1'
+    implementation 'com.google.maps.android:android-maps-utils:2.0.3'
 
     //testing
     testImplementation 'junit:junit:4.13'
