+++ lib/src/main/java/com/gmail/neooxpro/lib/network/directions/GoogleDirectionsApi.java
@@ -6,6 +6,7 @@ import io.reactivex.Single;
 import retrofit2.http.GET;
 import retrofit2.http.Query;
 
+@SuppressWarnings("PMD")
 public interface GoogleDirectionsApi {
     String GOOGLE_BASE_URL = "https://maps.googleapis.com/";
