+++ lib/src/main/java/com/gmail/neooxpro/lib/ui/view/ContactDetailsFragment.java
@@ -12,9 +12,6 @@ import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.ProgressBar;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
@@ -24,7 +21,6 @@ import androidx.core.content.ContextCompat;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentTransaction;
 
-
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.ViewModelProvider;
 
@@ -33,6 +29,7 @@ import com.gmail.neooxpro.lib.di.app.HasAppContainer;
 import com.gmail.neooxpro.lib.di.containers.ContactDetailsContainer;
 import com.gmail.neooxpro.lib.ui.FragmentListener;
 import com.gmail.neooxpro.lib.R;
+import com.gmail.neooxpro.lib.ui.delegates.ContactDetailsViewDelegate;
 import com.gmail.neooxpro.lib.ui.viewmodel.ContactDetailsViewModel;
 
 
@@ -40,18 +37,10 @@ import javax.inject.Inject;
 
 
 public class ContactDetailsFragment extends Fragment {
-    private TextView contactName;
-    private TextView contactPhone;
-    private TextView contactPhone2;
-    private TextView contactEmail1;
-    private TextView contactEmail2;
-    private TextView contactDescription;
-    private TextView contactBirthday;
-    private ProgressBar progressBar;
-    private Button birthButton;
     private static final int REQUEST_CODE = 1;
     private String contactId;
     private Toolbar toolbar;
+    private ContactDetailsViewDelegate viewDelegate;
 
     @Inject
     ViewModelProvider.Factory factory;
@@ -72,18 +61,10 @@ public class ContactDetailsFragment extends Fragment {
                              @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         setHasOptionsMenu(true);
         View view = inflater.inflate(R.layout.contact_details_fragment, container, false);
-        progressBar = view.findViewById(R.id.progress_bar_contactDetails);
+        viewDelegate = new ContactDetailsViewDelegate(view, getActivity(), model);
         toolbar.setTitle(R.string.contactDetails);
         assert this.getArguments() != null;
         contactId = this.getArguments().getString("args");
-        contactName = view.findViewById(R.id.contactName);
-        contactPhone = view.findViewById(R.id.contactPhone);
-        contactPhone2 = view.findViewById(R.id.contactPhone2);
-        contactEmail1 = view.findViewById(R.id.contactMail_1);
-        contactEmail2 = view.findViewById(R.id.contactMail_2);
-        contactDescription = view.findViewById(R.id.contactDescription);
-        contactBirthday = view.findViewById(R.id.contactBirthday);
-        birthButton = view.findViewById(R.id.birthDayButton);
 
         if (ContextCompat.checkSelfPermission(requireContext(),
                 Manifest.permission.READ_CONTACTS) != PackageManager.PERMISSION_GRANTED) {
@@ -108,36 +89,17 @@ public class ContactDetailsFragment extends Fragment {
     }
 
     public void queryContactDetails(@NonNull String id) {
-        LiveData<Boolean> progressBarStatus = model.isLoading();
+        viewDelegate.setProgressBar(getViewLifecycleOwner());
         LiveData<Contact> data = model.getData(id);
         model.haveNotification(id);
-        progressBarStatus.observe(getViewLifecycleOwner(),
-                isLoading -> progressBar.setVisibility(isLoading ? View.VISIBLE : View.GONE));
         data.observe(getViewLifecycleOwner(), contact -> {
-            contactName.setText(contact.getName());
-            contactPhone.setText(contact.getPhone());
-            contactPhone2.setText(contact.getPhone2());
-            contactEmail1.setText(contact.getEmail1());
-            contactEmail2.setText(contact.getEmail2());
-            contactDescription.setText(contact.getDescription());
-
-            contactBirthday.setText(String.format(getString(R.string.bTitle), contact.getBirthdayDate()));
-            if (contact.getBirthday() != null) {
-                notificationProcessing(contact);
-            } else {
-                birthButton.setText(R.string.noBirthdayDate);
-            }
+           viewDelegate.setViews(contact, getViewLifecycleOwner());
         });
     }
 
     public void onCreateOptionsMenu(@NonNull Menu menu, @NonNull MenuInflater inflater) {
         inflater.inflate(R.menu.menu_details, menu);
 
-    }
-
-    @NonNull
-    public ContactDetailsViewModel getModel() {
-        return model;
     }
 
     @Override
@@ -177,7 +139,7 @@ public class ContactDetailsFragment extends Fragment {
 
     private void openContactMapFragment() {
 
-        FragmentTransaction ft = getFragmentManager().beginTransaction();
+        FragmentTransaction ft = requireActivity().getSupportFragmentManager().beginTransaction();
         ContactMapFragment cdf = new ContactMapFragment();
         Bundle bundle = new Bundle();
         bundle.putString("args", contactId);
@@ -189,42 +151,11 @@ public class ContactDetailsFragment extends Fragment {
 
     }
 
-    public void notificationProcessing(@NonNull final Contact contact) {
-        LiveData<Boolean> haveNotification = model.getNotificationStatus();
-        haveNotification.observe(getViewLifecycleOwner(), this::setBirthButtonText);
-
-        birthButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-
-                getModel().enableOrDisableBirthdayNotification(contact);
-            }
-        });
-
-
-    }
-
-
-    private void setBirthButtonText(boolean haveNotification) {
-        if (haveNotification) {
-            birthButton.setText(R.string.notificationOn);
-        } else {
-            birthButton.setText(R.string.notificationOff);
-        }
-    }
-
     @Override
     public void onDestroyView() {
         super.onDestroyView();
-        contactName = null;
-        contactPhone = null;
-        contactPhone2 = null;
-        contactEmail1 = null;
-        contactEmail2 = null;
-        contactDescription = null;
-        contactBirthday = null;
-        birthButton = null;
-        progressBar = null;
+        viewDelegate.onDestroyView();
+
 
     }
