+++ lib/src/main/java/com/gmail/neooxpro/lib/ui/view/ContactListFragment.java
@@ -12,46 +12,35 @@ import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ProgressBar;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.SearchView;
 import androidx.appcompat.widget.Toolbar;
 import androidx.core.content.ContextCompat;
 import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentTransaction;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.ViewModelProvider;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
 
 import com.gmail.neooxpro.lib.di.app.HasAppContainer;
 import com.gmail.neooxpro.lib.di.containers.ContactListContainer;
 import com.gmail.neooxpro.lib.ui.FragmentListener;
-import com.gmail.neooxpro.lib.adapter.ContactItemDecoration;
-import com.gmail.neooxpro.lib.adapter.ContactsListAdapter;
 
 import com.gmail.neooxpro.java.domain.model.Contact;
 import com.gmail.neooxpro.lib.R;
+import com.gmail.neooxpro.lib.ui.delegates.ContactListViewDelegate;
 import com.gmail.neooxpro.lib.ui.viewmodel.ContactListViewModel;
 
 import java.util.ArrayList;
-import java.util.List;
 
 import javax.inject.Inject;
 
 
-public class ContactListFragment extends Fragment implements ContactsListAdapter.ItemClickListener {
+public class ContactListFragment extends Fragment {
     private static final int REQUEST_CODE = 1;
-    private static final int DP = 4;
     private Toolbar toolbar;
-    private RecyclerView recyclerView;
-    private ContactsListAdapter adapter;
-    private List<Contact> contactsList;
-    private ProgressBar progressBar;
+    private ContactListViewDelegate viewDelegate;
 
     @Inject
     ViewModelProvider.Factory factory;
@@ -82,19 +71,11 @@ public class ContactListFragment extends Fragment implements ContactsListAdapter
     }
 
     public void queryContacts() {
-        LiveData<Boolean> progressBarStatus = model.isLoading();
         LiveData<ArrayList<Contact>> data = model.getData();
         model.setSubject("");
-        progressBarStatus.observe(getViewLifecycleOwner(), isLoading -> {
-            if (isLoading != null) {
-                progressBar.setVisibility(isLoading ? View.VISIBLE : View.GONE);
-            }
-        });
+        viewDelegate.setProgressBar(getViewLifecycleOwner());
         data.observe(getViewLifecycleOwner(), contacts -> {
-            if (adapter != null) {
-                adapter.submitItems(contacts);
-                contactsList = contacts;
-            }
+            viewDelegate.setAdapter(contacts);
         });
 
     }
@@ -110,8 +91,8 @@ public class ContactListFragment extends Fragment implements ContactsListAdapter
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
         setHasOptionsMenu(true);
-        initView(view);
-        progressBar = view.findViewById(R.id.progress_bar_contactList);
+        viewDelegate = new ContactListViewDelegate(model, getActivity());
+        viewDelegate.initView(view);
         if (ContextCompat.checkSelfPermission(requireContext(),
                 Manifest.permission.READ_CONTACTS) != PackageManager.PERMISSION_GRANTED) {
             requestPermissions(new String[]{Manifest.permission.READ_CONTACTS}, REQUEST_CODE);
@@ -152,54 +133,17 @@ public class ContactListFragment extends Fragment implements ContactsListAdapter
     @Override
     public void onDestroyView() {
         super.onDestroyView();
-        adapter = null;
-        recyclerView = null;
-        progressBar = null;
+        viewDelegate.onDestroyView();
     }
 
     public void onCreateOptionsMenu(@NonNull Menu menu, @NonNull MenuInflater inflater) {
-        inflater.inflate(R.menu.menu, menu);
-        MenuItem searchItem = menu.findItem(R.id.action_search);
-        SearchView searchView = (SearchView) searchItem.getActionView();
-        searchView.setQueryHint(getString(R.string.search));
-        searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
-            @Override
-            public boolean onQueryTextSubmit(String query) {
-
-                return false;
-            }
-
-            @Override
-            public boolean onQueryTextChange(String newText) {
-                ContactListViewModel contactListViewModel = getModel();
-                if (contactListViewModel != null) {
-                    contactListViewModel.setSubject(newText);
-                }
-                return false;
-            }
-        });
-
+        viewDelegate.searchContact(menu, inflater);
         super.onCreateOptionsMenu(menu, inflater);
     }
 
     @Override
     public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-        if (item.getItemId() == R.id.map_item) {
-            openContactMapFragment();
-            return true;
-        }
-        return super.onOptionsItemSelected(item);
-    }
-
-    private void openContactMapFragment() {
-
-        FragmentTransaction ft = getFragmentManager().beginTransaction();
-        ContactListMapFragment cdf = new ContactListMapFragment();
-        ft
-                .replace(R.id.container, cdf)
-                .addToBackStack(null)
-                .commit();
-
+        return viewDelegate.onOptionsItemSelected(item) || super.onOptionsItemSelected(item);
     }
 
     @Override
@@ -209,33 +153,5 @@ public class ContactListFragment extends Fragment implements ContactsListAdapter
         ((AppCompatActivity) getActivity()).setSupportActionBar(toolbar);
     }
 
-    private void initView(@NonNull final View view) {
-        recyclerView = view.findViewById(R.id.contact_list_recycler_view);
-        recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
-        recyclerView.setHasFixedSize(true);
-        recyclerView.addItemDecoration(
-                new ContactItemDecoration(dpToPx(DP)));
-        adapter = new ContactsListAdapter();
-        adapter.setOnClickListener(this);
-        recyclerView.setAdapter(adapter);
-    }
-
-    private int dpToPx(int dp) {
-        return (int) (dp * getContext().getResources().getDisplayMetrics().density);
-    }
-
-    @Override
-    public void onItemClicked(int position) {
-
-        FragmentTransaction ft = getFragmentManager().beginTransaction();
-        ContactDetailsFragment cdf = new ContactDetailsFragment();
-        Bundle bundle = new Bundle();
-        bundle.putString("args", contactsList.get(position).getId());
-        cdf.setArguments(bundle);
-        ft
-                .replace(R.id.container, cdf)
-                .addToBackStack(null)
-                .commit();
-    }
 
 }
